Change "currLen" to uint32

+++ vlog/vlog.go
@@ -191,7 +191,7 @@ func (l *Log) At(loc item.Location, continueOnErr bool) Iter {
 	return Iter{
 		firstKey:      loc.Key,
 		currOff:       loc.Off,
-		currLen:       loc.Len,
+		currLen:       uint32(loc.Len),
 		continueOnErr: continueOnErr,
 	}
 }
